pyproject.toml
[project]
name = "alexa-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-health-checking>=1.74.0",
    "grpcio-tools>=1.74.0",
    "kokoro>=0.9.4",
    "ollama>=0.5.3",
    "openai-whisper>=20250625",
    "openwakeword>=0.4.0",
    "protobuf>=6.32.0",
    "psutil>=7.0.0",
    "torch>=2.8.0",
]
